requirements.txt
# base
Django==4.2.7
whitenoise==6.2.0
gunicorn==20.1.0
django-cors-headers==4.1.0
python-dotenv==1.0.1
requests==2.32.5

# posgresql
psycopg==3.2.3
psycopg2-binary==2.9.10

# images and s3
pillow==11.0.0
django-storages==1.13.2
boto3==1.26.137

# drf & jwt
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.4.0

# UI
django-jazzmin==3.0.1

# testing
selenium==4.20.0
beautifulsoup4==4.13.4
model-bakery==1.20.5
